+++ spring-batch-core/src/test/java/org/springframework/batch/core/launch/support/RunIdIncrementerTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2006-2010 the original author or authors.
+ * Copyright 2006-2013 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,7 @@
  */
 package org.springframework.batch.core.launch.support;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 import org.springframework.batch.core.JobParameters;
@@ -23,23 +23,24 @@ import org.springframework.batch.core.JobParametersBuilder;
 
 /**
  * @author Dave Syer
+ * @author Michael Minella
  *
  */
 public class RunIdIncrementerTests {
-	
+
 	private RunIdIncrementer incrementer = new RunIdIncrementer();
 
 	@Test
 	public void testGetNext() {
 		JobParameters next = incrementer.getNext(null);
-		assertEquals(1, next.getLong("run.id"));
-		assertEquals(2, incrementer.getNext(next).getLong("run.id"));
+		assertEquals(1, next.getLong("run.id").intValue());
+		assertEquals(2, incrementer.getNext(next).getLong("run.id").intValue());
 	}
 
 	@Test
 	public void testGetNextAppends() {
 		JobParameters next = incrementer.getNext(new JobParametersBuilder().addString("foo", "bar").toJobParameters());
-		assertEquals(1, next.getLong("run.id"));
+		assertEquals(1, next.getLong("run.id").intValue());
 		assertEquals("bar", next.getString("foo"));
 	}
 
@@ -47,7 +48,7 @@ public class RunIdIncrementerTests {
 	public void testGetNextNamed() {
 		incrementer.setKey("foo");
 		JobParameters next = incrementer.getNext(null);
-		assertEquals(1, next.getLong("foo"));
+		assertEquals(1, next.getLong("foo").intValue());
 	}
 
 }
